# swagger.yaml
openapi: 3.0.0
info:
  title: Adhivasindo Backend API
  version: 1.0.0
  description: |
    Take Home Test Backend Developer — Adhivasindo.

    **Cara pakai:**
    1. Jalankan `POST /api/auth/login` dengan `admin` / `admin123`
    2. Copy token dari response
    3. Klik tombol **Authorize** di kanan atas, paste token (tanpa kata "Bearer")
    4. Endpoint lain sudah otomatis terautentikasi
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/users:
    get:
      summary: "Endpoint B — Get all users"
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Daftar user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  total: { type: integer, example: 2 }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: "Endpoint B — Create user"
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string, example: budi }
                email: { type: string, example: budi@mail.com }
                password: { type: string, example: rahasia123 }
                full_name: { type: string, example: "Budi Santoso" }
      responses:
        "201":
          description: User dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  data:
                    $ref: "#/components/schemas/User"
        "409":
          description: Username/email sudah ada
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Masukkan JWT token dari /api/auth/login (tanpa kata 'Bearer')"
  schemas:
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        username: { type: string, example: admin }
        email: { type: string, example: "admin@adhivasindo.co.id" }
        full_name: { type: string, example: Administrator }
        is_active: { type: boolean, example: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Invalid credentials." }
        error: { type: string }
  responses:
    Unauthorized:
      description: Token tidak ada / tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"